Prevent errors if `destroy()` is called twice
Add checks before accessing object properties to see if they hadn't already been set to `null`.

Not modified the package `org.flixel.system.debug.*`

+++ org/flixel/system/FlxQuadTree.as
@@ -258,13 +258,17 @@ package org.flixel.system
 		 */
 		public function destroy():void
 		{
-			_headA.destroy();
+			if(_headA != null)
+				_headA.destroy();
 			_headA = null;
-			_tailA.destroy();
+			if(_tailA != null)
+				_tailA.destroy();
 			_tailA = null;
-			_headB.destroy();
+			if(_headB != null)
+				_headB.destroy();
 			_headB = null;
-			_tailB.destroy();
+			if(_tailB != null)
+				_tailB.destroy();
 			_tailB = null;
 
 			if(_northWestTree != null)
